Connectivité (#76)
* Début de la recherche pour connectivité réseau

* Test d'une méthode, mais warnings

* Before rebase

* Connectivity test before query; Need to be verify on a real IOS device

* Suppression de lignes non utilisées

* meilleur gestion de la connectivité , à tester sur device iOS

* hotfix pour cacher un composant

* cache de l'icone

+++ src/WS/AppelWebService.js
@@ -1,6 +1,44 @@
+import { NetInfo } from 'react-native';
+
 import { urlWs } from '../Util';
+import { showInformativePopin } from '../Popin';
+import { errorPopinTitle, noNetwork } from '../errors/ErrorStrings';
 import { tokenIsValid } from '../errors/Token';
 
+export const NO_CONNECTION = 'NO_CONNECTION';
+let isConnected = false;
+/**
+ * Handling network connection
+ */
+export async function isNetworkConnected()
+{
+  NetInfo.isConnected.addEventListener(
+    'change',
+    (listenNetwork) =>
+    {
+      if (!listenNetwork)
+      {
+        isConnected = false;
+      }
+      else
+      {
+        isConnected = true;
+      }
+    });
+  NetInfo.isConnected.fetch().then((listenNetwork) =>
+  {
+    if (!listenNetwork)
+    {
+      isConnected = false;
+      showInformativePopin(errorPopinTitle, noNetwork);
+    }
+    else
+    {
+      isConnected = true;
+    }
+  });
+  return isConnected;
+}
 
 async function getResponseObject(webResponse)
 {
@@ -37,6 +75,13 @@ async function verifToken(response, token, propsNavigation)
 
 export async function appelGet(lien, token, propsNavigation)
 {
+  // No connection => service not called
+  const connected = await isNetworkConnected();
+  if (connected === false)
+  {
+    return Promise.reject(NO_CONNECTION);
+  }
+
   const response = await fetch(`${urlWs}${lien}`, {
     method: 'GET',
     headers: {
@@ -50,6 +95,13 @@ export async function appelGet(lien, token, propsNavigation)
 
 export async function appelPost(lien, data, token, propsNavigation)
 {
+  // No connection => service not called
+  const connected = await isNetworkConnected();
+  if (connected === false)
+  {
+    return Promise.reject(NO_CONNECTION);
+  }
+
   const response = await fetch(`${urlWs}${lien}`, {
     method: 'POST',
     headers: {
@@ -65,6 +117,13 @@ export async function appelPost(lien, data, token, propsNavigation)
 
 export async function appelDelete(lien, token, propsNavigation)
 {
+  // No connection => service not called
+  const connected = await isNetworkConnected();
+  if (connected === false)
+  {
+    return Promise.reject(NO_CONNECTION);
+  }
+
   const response = await fetch(`${urlWs}${lien}`, {
     method: 'DELETE',
     headers: {
@@ -78,6 +137,13 @@ export async function appelDelete(lien, token, propsNavigation)
 
 export async function appelPut(lien, data, token, propsNavigation)
 {
+  // No connection => service not called
+  const connected = await isNetworkConnected();
+  if (connected === false)
+  {
+    return Promise.reject(NO_CONNECTION);
+  }
+
   const response = await fetch(`${urlWs}${lien}`, {
     method: 'PUT',
     headers: {
